finish with 2nd tester

+++ ft_strlen.c
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (*s)
@@ -24,3 +24,4 @@ size_t	ft_strlen(const char *s)
 	}
 	return (len);
 }
+
